+++ src/GeoFormats/GeoFormats.csproj
@@ -2,6 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>netcoreapp2.0</TargetFramework>
+    <RootNamespace>Geo.Formats</RootNamespace>
   </PropertyGroup>
 
 </Project>
