+++ src/GeoOptions/GeoOptions.csproj
@@ -2,7 +2,13 @@
 
   <PropertyGroup>
     <TargetFramework>netcoreapp2.0</TargetFramework>
+    <RootNamespace>Geo.Options</RootNamespace>
   </PropertyGroup>
+
+  <ItemGroup>
+    <Compile Remove="Keys.cs" />
+  </ItemGroup>
+
 
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Configuration" Version="2.0.0" />
