+++ src/GeoOptions/Keys.cs
@@ -13,7 +13,6 @@
 
 	public class NokiaApiOptions
 	{
-		public string Key { get; set; }
-		public string Id { get; set; }
+		
 	}
 }
