--- src/GeoTests2/BingTests.cs
@@ -1,12 +1,0 @@
-﻿using Xunit;
-
-namespace Geo.Tests
-{
-  public static class BingTests
-  {
-    [Fact]
-    public static void FindFortWorth()
-    {
-    }
-  }
-}
