--- src/GeoTests2/GoogleTests.cs
@@ -1,14 +1,0 @@
-﻿
-
-using Xunit;
-
-namespace Geo.Tests
-{
-  public static class GoogleTests
-  {
-    [Fact]
-    public static void FindFortWorth()
-    {
-    }
-  }
-}
