--- src/GeoTests2/LocationMaster.cs
@@ -1,7 +1,0 @@
-﻿namespace Geo.Tests
-{
-  public static class LocationMaster
-  {
-    public const string FortWorth = "{\"Latitude\":32.75095,\"Longitude\":-97.33086}";
-  }
-}
