--- src/GeoTests2/MapQuestTests.cs
@@ -1,12 +1,0 @@
-﻿using Xunit;
-
-namespace Geo.Tests
-{
-  public static class MapQuestTests
-  {
-    [Fact]
-    public static void FindFortWorth()
-    {
-    }
-  }
-}
