--- src/GeoTests2/NokiaTests.cs
@@ -1,12 +1,0 @@
-﻿using Xunit;
-
-namespace Geo.Tests
-{
-  public static class NokiaTests
-  {
-    [Fact]
-    public static void FindFortWorth()
-    {
-    }
-  }
-}
