+++ src/Geocoder/Geocoder.csproj
@@ -2,6 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>netcoreapp2.0</TargetFramework>
+    <RootNamespace>Geo.Coder</RootNamespace>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Newtonsoft.Json" Version="10.0.3" />
@@ -9,8 +10,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\GeoAPIKeys\GeoKeys.csproj" />
     <ProjectReference Include="..\GeoFormats\GeoFormats.csproj" />
+    <ProjectReference Include="..\GeoOptions\GeoOptions.csproj" />
   </ItemGroup>
 
 </Project>
